requirements.txt
# Core
pymupdf==1.24.9
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4

# Embeddings / Vector index / Rerank
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
scikit-learn==1.5.1
torch>=2.1.0

# API / UI
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
streamlit==1.37.0

# Utilities
tqdm==4.66.4
orjson==3.10.7
